optimized splitting

+++ examples/SoftLed_01/SoftLed_01.ino
@@ -1,15 +1,15 @@
 #include <MobaTools.h>
 /* Demo: weiches Auf/Abblenden einer LED
  *  Mit der Klasse SoftLed lässt sich eine Led 'weich' ein- und ausschalten
- * Die Led muss an einem PWM-fähigen pin angeschlossen sein, d.h. an einem Pin
- * an dem auch analogWrite() funktioniert.
+ * Die Led muss nicht an einem PWM-fähigen Pin angeschlossen sein, es funktioniert
+ * mit jedem Digital-Pin.
  * SoftLed.attach( byte Pinnr );// Anschlußpin für die Led festlegen
  * SoftLed.riseTime( int zeit );// Zeit für auf/abblenden in ms
  * SoftLed.on();                // 'weich' einschalten
  * SoftLed.off();               // 'weich' ausschalten
 */
 
-const int ledPin =  5;  // PWM - fähiger Pin
+const byte ledPin =  5;  // beliebiger Digital Pin
 
 
 SoftLed meinLed;
